refactor: A LOT of cleanup
