--- lib/digitalocean.rb
@@ -1,4 +1,0 @@
-require_relative './digitalocean/upsert_record'
-
-module Digitalocean
-end
