+++ package.json
@@ -35,10 +35,11 @@
     "start": "PORT=3006 react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "test:utils": "bundle exec rake test",
     "eject": "react-scripts eject",
     "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
     "format:check": "prettier --check \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
-    "deploy:review": "bin/thor cli:deploy"
+    "deploy:review": "bin/thor story-cli:app:deploy"
   },
   "jest": {
     "transformIgnorePatterns": [
